ui: factor out com/leftnav and reuse across pages
