+++ ui/less/com/leftnav.less
@@ -1,0 +1,25 @@
+.leftnav {
+  flex: 0 0 175px;
+  
+  .leftnav-heading {
+    margin: 15px 0 5px 10px;
+    font-weight: bold;
+    color: gray;
+    font-size: 12px;
+    text-transform: uppercase;
+  }
+
+  .leftnav-link {
+    margin: 5px 0 5px 10px;
+    a {
+      color: #555;
+      &.selected {
+        color: #006EFF;
+      }
+    }
+    i { 
+      width: 20px;
+      text-align: center;
+    }
+  }
+}
