+++ ui/less/main.less
@@ -44,11 +44,11 @@
 @import "com/channel-list.less";
 @import "com/users/follow-list.less";
 @import "com/findbar.less";
+@import "com/leftnav.less";
 
 @import "transitions/slideup.less";
 @import "transitions/fade.less";
 
-@import "views/leftnav.less";
 @import "views/rightnav.less";
 @import "views/newsfeed.less";
 @import "views/inbox.less";
