--- ui/less/views/leftnav.less
@@ -1,5 +1,0 @@
-#leftnav {
-  width: 200px;
-  background: #fff;
-  border-right: 1px solid #ccc;
-}
