+++ ui/less/views/newsfeed.less
@@ -1,31 +1,6 @@
 #newsfeed {
   display: flex;
-
-  .newsfeed-leftnav {
-    flex: 0 0 175px;
-    
-    .newsfeed-leftnav-heading {
-      margin: 15px 0 5px 10px;
-      font-weight: bold;
-      color: gray;
-      font-size: 12px;
-      text-transform: uppercase;
-    }
-
-    .newsfeed-leftnav-link {
-      margin: 5px 0 5px 10px;
-      a {
-        color: #555;
-        &.selected {
-          color: #006EFF;
-        }
-      }
-      i { 
-        width: 20px;
-        text-align: center;
-      }
-    }
-  }
+  
   .newsfeed-rightnav {
     flex: 0 0 175px;
     font-weight: 300;
@@ -75,7 +50,6 @@
   }
 
   .msg-list-items {
-    padding-right: 5px;
     hr {
       max-width: 660px;
       margin-left: auto;
