+++ ui/views/inbox.jsx
@@ -2,6 +2,7 @@
 import React from 'react'
 import pull from 'pull-stream'
 import mlib from 'ssb-msgs'
+import LeftNav from '../com/leftnav'
 import MsgList from '../com/msg-list'
 import Summary from '../com/msg-view/summary'
 import * as HelpCards from '../com/help/cards'
@@ -29,13 +30,6 @@ export default class Inbox extends React.Component {
   }
 
   render() {
-    const Hero = props => {
-      return <div className="hero">
-        <h1>Inbox</h1>
-        <div>Private, encrypted messages.</div>
-      </div>
-    }
-
     return <div id="inbox">
       <MsgList
         ref="list"
@@ -43,7 +37,7 @@ export default class Inbox extends React.Component {
         dateDividers
         composer composerProps={{placeholder: 'Write a new private message'}}
         ListItem={Summary} listItemProps={{ userPic: true }}
-        Hero={Hero}
+        LeftNav={LeftNav} leftNavProps={{ location: this.props.location }}
         live={{ gt: [Date.now(), null] }}
         emptyMsg="Your inbox is empty."
         append={this.helpCards.bind(this)}
